+++ src/mathengine/vector/Vector2D.java
@@ -1,10 +1,10 @@
 package mathengine.vector;
 
 public class Vector2D extends Vector {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public Vector2D(float x, float y) {
+    public Vector2D(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -14,40 +14,40 @@ public class Vector2D extends Vector {
         this.y = vector.getY();
     }
 
-    public final float getX() {
+    public final double getX() {
         return this.x;
     }
 
-    public final float getY() {
+    public final double getY() {
         return this.y;
     }
 
-    public final float getZ() {
+    public final double getZ() {
         return 0;
     }
 
-    public Vector2D setX(float x) {
+    public Vector2D setX(double x) {
         this.x = x;
         return this;
     }
 
-    public Vector2D setY(float y) {
+    public Vector2D setY(double y) {
         this.y = y;
         return this;
     }
 
     @Override
-    protected final Vector setNew(float x, float y, float z) {
+    protected final Vector setNew(double x, double y, double z) {
         return new Vector2D(x, y);
     }
 
-    public void set(float x, float y, float z) {
-        x = Float.isNaN(x) ? 0 : this.x;
-        y = Float.isNaN(y) ? 0 : this.y;
+    public void set(double x, double y, double z) {
+        x = Double.isNaN(x) ? 0 : this.x;
+        y = Double.isNaN(y) ? 0 : this.y;
         this.setX(x).setY(y);
     }
 
-    public void set(float x, float y) {
+    public void set(double x, double y) {
         this.set(x, y, this.getZ());
     }
 
@@ -59,11 +59,11 @@ public class Vector2D extends Vector {
         return new Vector2D(1, 1);
     }
 
-    public Vector sum(float x, float y) {
+    public Vector sum(double x, double y) {
         return this.sum(x, y, 0);
     }
 
-    public Vector subtract(float x, float y) {
+    public Vector subtract(double x, double y) {
         return this.subtract(x, y, 0);
     }
 }
